foodgram/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.4
black==20.8b1
click==7.1.2
distlib==0.3.1
Django==3.1.7
django-debug-toolbar==3.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
flake8==3.9.1
gunicorn==20.1.0
isort==5.8.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.7.4
python-dotenv==0.17.0
pytz==2021.1
regex==2021.3.17
reportlab==3.5.67
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.3
wrapt==1.12.1
